+++ src/components/navigation-accordion/navigation-accordion.js
@@ -77,6 +77,11 @@ class NavigationAccordion extends React.PureComponent {
               <li key={subItem.path} className="mt6">
                 <a href={`#${subItem.path}`} className={itemClasses(isActive)}>
                   {subItem.title}
+                  {subItem.tag ? (
+                    <span className="ml6">{subItem.tag}</span>
+                  ) : (
+                    ''
+                  )}
                 </a>
               </li>
             );
@@ -85,6 +90,7 @@ class NavigationAccordion extends React.PureComponent {
           <li key={item.path} className="mb6">
             <a href={`#${item.path}`} className={itemClasses(isActive)}>
               {item.title}
+              {item.tag ? <span className="ml6">{item.tag}</span> : ''}
             </a>
             <ul className={openSubItems ? 'pl12 color-darken75' : 'none'}>
               {subItems}
@@ -136,7 +142,7 @@ class NavigationAccordion extends React.PureComponent {
               >
                 <div className={textClasses}>
                   {title}
-                  {page.tag ? page.tag : ''}
+                  {page.tag ? <span className="ml6">{page.tag}</span> : ''}
                 </div>
                 {icon}
               </a>
@@ -174,10 +180,12 @@ NavigationAccordion.propTypes = {
     secondLevelItems: PropTypes.arrayOf(
       PropTypes.shape({
         title: PropTypes.string.isRequired,
+        tag: PropTypes.node,
         path: PropTypes.string.isRequired,
         thirdLevelItems: PropTypes.arrayOf(
           PropTypes.shape({
             title: PropTypes.string.isRequired,
+            tag: PropTypes.node,
             path: PropTypes.string.isRequired
           })
         )